+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog
 
 [Unreleased]
+- Added Honduras to countries list
 
 ### [2.19.7] - 2022-07-12
 - Update headers when processing pending transactions